# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Produk API",
    "description": "Backend CRUD untuk tabel produk di database Neon.",
    "version": "1.0.0"
  },
  "paths": {
    "/": {
      "get": {
        "summary": "Root",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/produk": {
      "get": {
        "summary": "Ambil daftar produk",
        "operationId": "list_produk_produk_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Produk"
                  },
                  "type": "array",
                  "title": "Response List Produk Produk Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Tambah produk",
        "operationId": "create_produk_produk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProdukInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Produk"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/produk/{produk_id}": {
      "get": {
        "summary": "Ambil satu produk",
        "operationId": "get_produk_produk__produk_id__get",
        "parameters": [
          {
            "name": "produk_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Produk Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Produk"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Ubah produk",
        "operationId": "update_produk_produk__produk_id__put",
        "parameters": [
          {
            "name": "produk_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Produk Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProdukInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Produk"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Hapus produk",
        "operationId": "delete_produk_produk__produk_id__delete",
        "parameters": [
          {
            "name": "produk_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Produk Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Produk": {
        "properties": {
          "nama": {
            "type": "string",
            "title": "Nama"
          },
          "harga": {
            "type": "integer",
            "title": "Harga"
          },
          "stok": {
            "type": "integer",
            "title": "Stok",
            "default": 0
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "nama",
          "harga",
          "id"
        ],
        "title": "Produk"
      },
      "ProdukInput": {
        "properties": {
          "nama": {
            "type": "string",
            "title": "Nama"
          },
          "harga": {
            "type": "integer",
            "title": "Harga"
          },
          "stok": {
            "type": "integer",
            "title": "Stok",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "nama",
          "harga"
        ],
        "title": "ProdukInput"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}